requirements.txt
langchain_google_genai
langchain_core
langchain_community
langchain_chroma
langchain
python-dotenv
pydantic
SQLAlchemy
typing_extensions
langgraph
onnxruntime
pypdf